# Web/packages/api/openapi/a-health-api-v1.yaml
openapi: 3.1.0
info:
  title: A-health Platform API
  version: 1.0.0-phase1
  summary: Contract for Phase 0 (Identity) and Phase 1 (Core Consultation).
  description: |
    Authoritative contract for the A-health platform. Frontend and backend build
    against this document in parallel; it is the single source of truth for
    request/response shapes.

    ## Conventions

    - REST over HTTPS. JSON only. Base path `/api/v1`.
    - Field naming: `snake_case` everywhere, no exceptions.
    - All timestamps are RFC 3339 / ISO 8601 in UTC with a `Z` suffix.
    - All identifiers are UUID v4 strings.
    - Every mutable resource carries `version` (integer) and `updated_at`.
      Clients send `base_version` when writing to detect conflicts.

    ## Authentication

    Bearer JWT. Access token TTL 15 minutes, refresh token TTL 30 days and
    rotated on every use. RBAC is enforced at the gateway AND re-validated in
    the service layer — never rely on the client hiding a control.

    ## State transitions

    Clinical state changes are expressed as action sub-resources
    (`POST /consultations/{id}/accept`), never as a `PATCH` on a status field.
    This pattern is applied uniformly across every resource so there is one rule
    to remember. Each action has its own permission check, its own audit entry,
    and its own set of legal source states.

    ## Idempotency

    Every POST that creates or transitions clinical state requires an
...

    ## Pagination

...

...

    ## Realtime

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/register/patient:
    post:
      tags: [auth]
      summary: Register a patient and trigger an OTP
      description: |
        Phone number is the primary identifier. No password is set for patient
        accounts — authentication is OTP only. Returns 202 because the account
        is not usable until the OTP is verified.

        Registering an already-registered number returns 202 as well, with an
        OTP sent, so the endpoint cannot be used to enumerate which numbers
        exist on the platform.
      operationId: registerPatient
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number, preferred_language]
              properties:
                phone_number:
                  $ref: '#/components/schemas/PhoneNumber'
                full_name:
                  type: string
                  maxLength: 150
                preferred_language:
                  $ref: '#/components/schemas/LanguageCode'
                channel:
                  $ref: '#/components/schemas/Channel'
      responses:
        '202':
          description: OTP dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpChallenge'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register/clinician:
    post:
      tags: [auth]
      summary: Register a clinician pending verification
      description: |
        Creates an account with `status = pending_verification`. The clinician
        cannot accept consultations, and no consultation note can be signed by
        them, until an admin approves the licence via
        `POST /clinicians/{clinician_id}/verification`.

        `verification_document_keys` are object-storage keys obtained from a
        prior upload; raw files are never posted to this endpoint.
      operationId: registerClinician
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [phone_number, full_name, email, license_number, specialty, preferred_language]
              properties:
                phone_number:
                  $ref: '#/components/schemas/PhoneNumber'
                full_name:
                  type: string
                  maxLength: 150
                email:
                  type: string
                  format: email
                license_number:
                  type: string
                  maxLength: 50
                specialty:
                  $ref: '#/components/schemas/Specialty'
                languages_spoken:
                  type: array
                  items:
                    $ref: '#/components/schemas/LanguageCode'
                preferred_language:
                  $ref: '#/components/schemas/LanguageCode'
                facility_id:
                  type: string
                  format: uuid
                  description: Optional. Reference only — does not grant facility privileges.
                verification_document_keys:
                  type: array
                  items: { type: string }
      responses:
        '202':
          description: Registered, verification pending, OTP dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpChallenge'
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/otp/request:
    post:
      tags: [auth]
      summary: Request an OTP for an existing account
      description: |
        Rate limited per phone number and per source IP. Always returns 202
        whether or not the number is registered, to prevent enumeration.
      operationId: requestOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number]
              properties:
                phone_number:
                  $ref: '#/components/schemas/PhoneNumber'
                channel:
                  $ref: '#/components/schemas/Channel'
      responses:
        '202':
          description: OTP dispatched if the account exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OtpChallenge'
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify an OTP and issue tokens
      description: |
        Six digits, five-minute expiry, maximum five attempts per challenge.
        Exceeding the attempt limit invalidates the challenge entirely — a new
        one must be requested.
      operationId: verifyOtp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                challenge_id:
                  type: string
                  format: uuid
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
                device_id:
                  type: string
                  description: Stable client-generated device identifier, for session management.
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /users/me:
    get:
      tags: [users]
      summary: Get the authenticated user's profile
      operationId: getMe
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthenticated' }
    patch:
      tags: [users]
      summary: Update own profile
      description: |
        Non-clinical fields only. Clinical history is never edited through this
        endpoint — it is written by clinicians through consultation notes.
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [base_version]
              properties:
                base_version:
                  type: integer
                full_name:
                  type: string
                  maxLength: 150
                preferred_language:
                  $ref: '#/components/schemas/LanguageCode'
                default_location:
                  $ref: '#/components/schemas/GeoPoint'
                region_code:
                  type: string
                  maxLength: 20
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
components:
  securitySchemes:
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: |
        HMAC of the raw request body using the shared per-provider secret.
        Used only on /payments/webhooks/*, paired with a source-IP allow-list
        where the provider publishes one.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    gatewaySignature:
      type: apiKey
      in: header
      name: X-Gateway-Signature
      description: |
        HMAC of the raw request body using the shared aggregator secret.
        Used only on `/gateway/*`. Paired with a source-IP allow-list.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUID, created at the moment the user acted. Reused
        unchanged across every retry of the same intent. The server replays the
        original response for 24 hours.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthenticated:
      description: Missing, expired, or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Conflict:
      description: State conflict or version mismatch
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnprocessableEntity:
      description: Well-formed but semantically invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Human-readable, already localised to the caller's `Accept-Language`.
            field:
              type: string
              description: Present on validation errors, naming the offending field.
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              description: Echoes `X-Request-ID`. Quote this when reporting a problem.
    ErrorCode:
      type: string
      description: |
        Closed catalogue. Clients branch on `code`, never on `message`.
        Adding a code is a non-breaking change; removing or repurposing one is
        breaking and requires a new API version.
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - TOKEN_INVALID
        - TOKEN_REUSE_DETECTED
        - OTP_INVALID
        - OTP_EXPIRED
        - OTP_ATTEMPTS_EXCEEDED
        - FORBIDDEN
        - ROLE_NOT_PERMITTED
        - NOT_RESOURCE_OWNER
        - NOT_FOUND
        - IDEMPOTENCY_KEY_CONFLICT
        - VERSION_CONFLICT
        - STATE_TRANSITION_INVALID
        - CLINICIAN_NOT_VERIFIED
        - CLINICIAN_UNAVAILABLE
        - CONSULTATION_ALREADY_ASSIGNED
        - CARE_THREAD_CLOSED
        - SLOT_UNAVAILABLE
        - SLOT_IN_PAST
        - APPOINTMENT_NOT_YET_STARTABLE
        - APPOINTMENT_ALREADY_STARTED
        - CHECK_IN_ALREADY_ANSWERED
        - FOLLOW_UP_CYCLE_CLOSED
        - PRESCRIPTION_NOT_ACTIVE
        - CONSENT_REQUIRED
        - DEVICE_REVOKED
        - DISPENSE_CODE_INVALID
        - DISPENSE_CODE_EXPIRED
        - RESULT_ALREADY_ACKNOWLEDGED
        - ETHICS_APPROVAL_REQUIRED
        - AGGREGATE_TOO_SMALL
        - FACILITY_NOT_INTEGRATED
        - ALREADY_RATED
        - PAYMENT_METHOD_UNAVAILABLE
        - PAYMENT_ALREADY_SETTLED
        - REFUND_EXCEEDS_PAYMENT
        - PAYER_PHONE_REQUIRED
        - DUPLICATE_RESOURCE
        - RATE_LIMITED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
...
    PhoneNumber:
      type: string
      description: E.164 format, including country code.
      pattern: '^\+[1-9][0-9]{7,14}$'
      examples: ['+255712345678']
    LanguageCode:
      type: string
      description: ISO 639-1.
      enum: [sw, en, fr, ha, am]
    Channel:
      type: string
      enum: [app, sms, ussd, voice, web]
    Sex:
      type: string
      enum: [male, female, other]
    Specialty:
      type: string
      enum:
        - general_practice
        - internal_medicine
        - paediatrics
        - obstetrics_gynaecology
        - surgery
        - oncology
        - psychiatry
        - dermatology
        - other
...
    UserRole:
      type: string
      enum: [patient, clinician, dispatcher, facility_admin, platform_admin]
    UserStatus:
      type: string
      enum: [active, suspended, pending_verification, deactivated]
    VerificationStatus:
      type: string
      enum: [pending, verified, rejected]
    GeoPoint:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        accuracy_metres: { type: number }
    OtpChallenge:
      type: object
      required: [challenge_id, expires_at, delivery_channel]
      properties:
        challenge_id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        delivery_channel:
          $ref: '#/components/schemas/Channel'
        masked_destination:
          type: string
          description: Partially redacted destination, e.g. `+2557****5678`.
          examples: ['+2557****5678']
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in, token_type]
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime in seconds.
          examples: [900]
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, role, status, preferred_language, version, updated_at]
      properties:
        id: { type: string, format: uuid }
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        full_name: { type: string }
        phone_number:
          $ref: '#/components/schemas/PhoneNumber'
        email: { type: [string, 'null'], format: email }
        preferred_language:
          $ref: '#/components/schemas/LanguageCode'
        patient_profile:
          $ref: '#/components/schemas/PatientProfile'
        clinician_profile:
          $ref: '#/components/schemas/ClinicianProfile'
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PatientProfile:
      type: object
      required: [id, version, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: [string, 'null'], format: uuid }
        guardian_user_id:
          type: [string, 'null']
          format: uuid
          description: Set for dependant profiles that have no login of their own.
        full_name: { type: string }
        date_of_birth: { type: string, format: date }
        sex:
          $ref: '#/components/schemas/Sex'
        chronic_conditions:
          type: array
          items: { type: string }
        allergies:
          type: array
          items: { type: string }
        blood_type: { type: [string, 'null'] }
        default_location:
          $ref: '#/components/schemas/GeoPoint'
        region_code:
          type: string
          description: Used for de-identified aggregation. Never returned in analytics views alongside an id.
        version: { type: integer }
        updated_at: { type: string, format: date-time }
    ClinicianProfile:
      type: object
      required: [id, verification_status, version, updated_at]
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        facility_id: { type: [string, 'null'], format: uuid }
        license_number: { type: string }
        specialty:
          $ref: '#/components/schemas/Specialty'
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        rejection_reason: { type: [string, 'null'] }
        languages_spoken:
          type: array
          items:
            $ref: '#/components/schemas/LanguageCode'
        is_available: { type: boolean }
        current_load:
          type: integer
          description: Count of consultations currently in progress. Read-only.
        rating_avg: { type: [number, 'null'], format: float }
        version: { type: integer }
        updated_at: { type: string, format: date-time }
...